Unit tests for cloning config data
