+++ src/config/git_config_test.go
@@ -7,35 +7,22 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGitConfigCache(t *testing.T) {
+func TestGitConfig(t *testing.T) {
 	t.Parallel()
 	t.Run("Clone", func(t *testing.T) {
 		t.Parallel()
-		alpha := config.Key{"alpha"}
-		beta := config.Key{"beta"}
-		original := config.GitConfigCache{
-			alpha: "A",
-			beta:  "B",
+		original := config.GitConfig{
+			Global: config.GitConfigCache{
+				config.KeyOffline: "1",
+			},
+			Local: config.GitConfigCache{
+				config.KeyMainBranch: "main",
+			},
 		}
-		cloned := original.Clone()
-		cloned[alpha] = "new A"
-		cloned[beta] = "new B"
-		assert.Equal(t, "A", original[alpha])
-		assert.Equal(t, "B", original[beta])
-	})
-
-	t.Run("KeysMatching", func(t *testing.T) {
-		t.Parallel()
-		cache := config.GitConfigCache{
-			config.Key{"key1"}:  "A",
-			config.Key{"key2"}:  "B",
-			config.Key{"other"}: "other",
-		}
-		have := cache.KeysMatching("key")
-		want := []config.Key{
-			{"key1"},
-			{"key2"},
-		}
-		assert.Equal(t, want, have)
+		clone := original.Clone()
+		clone.Global[config.KeyOffline] = "0"
+		clone.Local[config.KeyMainBranch] = "dev"
+		assert.Equal(t, "1", original.Global[config.KeyOffline])
+		assert.Equal(t, "main", original.Local[config.KeyMainBranch])
 	})
 }
